Refactored such that all the business logic from scripts/sc-configurable.py is now in src/config.py, primarily but not entirely in a class called Configurator
